requirements.txt
absl-py==1.0.0
adjustText==0.7.3
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
beautifulsoup4==4.10.0
blis==0.7.5
bpemb==0.3.3
brotlipy==0.7.0
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
conllu==4.4.1
contextvars==2.4
cryptography==36.0.0
cycler==0.11.0
cymem==2.0.6
datasets==1.18.3
Deprecated==1.2.13
dill==0.3.4
dpcpp-cpp-rt==2022.1.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.2.0/en_core_web_md-3.2.0-py3-none-any.whl
executing==0.8.3
fastapi==0.76.0
filelock==3.4.1
fonttools==4.33.3
frozenlist==1.2.0
fsspec==2022.1.0
ftfy==6.0.3
gdown==4.3.0
gensim==4.1.2
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.0
huggingface-hub==0.4.0
icecream==2.1.2
idna==3.3
idna-ssl==1.1.0
immutables==0.16
importlib-metadata==3.10.1
intel-cmplr-lib-rt==2022.1.0
intel-cmplr-lic-rt==2022.1.0
intel-opencl-rt==2022.1.0
intel-openmp==2022.1.0
Janome==0.4.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.1
konoha==4.6.5
langcodes==3.3.0
langdetect==1.0.9
lxml==4.7.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.2
mkl==2022.1.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==8.8.0
mpld3==0.3
multidict==5.2.0
multiprocess==0.70.12.2
murmurhash==1.0.6
nltk==3.6.7
numpy==1.22.3
oauthlib==3.2.0
overrides==3.1.0
packaging==21.3
pandas==1.1.5
pathy==0.6.1
Pillow==9.0.1
pip==21.2.4
preshed==3.0.6
protobuf==3.20.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.5.4
seaborn==0.11.2
segtok==1.5.11
sentence-transformers==2.2.0
sentencepiece==0.1.95
setuptools==61.2.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.1
spacy==3.2.2
spacy-legacy==3.0.9
spacy-loggers==1.0.1
sqlitedict==1.7.0
srsly==2.4.2
starlette==0.18.0
stefutils==0.15.0
sty==1.0.4
tabulate==0.8.9
tbb==2021.6.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thinc==8.0.13
threadpoolctl==3.1.0
tokenizers==0.11.4
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tqdm==4.62.3
transformers==4.16.2
tsnecuda==3.0.1
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.9
wasabi==0.9.0
wcwidth==0.2.5
Werkzeug==2.1.2
wheel==0.37.1
Wikipedia-API==0.5.4
wordninja==2.0.0
wrapt==1.13.3
xxhash==2.0.2
yarl==1.7.2
zipp==3.6.0
